pyproject.toml
[project]
authors = [
    {name = "Akhil Narang", email = "me@akhilnarang.dev"},
]
requires-python = "<4.0,>=3.10"
dependencies = [
    "python-telegram-bot[job-queue]>=21.0,<22.0",
    "pydantic<3.0.0,>=2.5.2",
    "httpx<1.0.0,>=0.25.2",
    "rich<14.0.0,>=13.7.0",
    "pydantic-settings<3.0.0,>=2.1.0",
    "redis<6.0.0,>=5.0.0",
    "arq>=0.26.0,<1.0.0",
    "google-generativeai>=0.8.0,<1.0.0",
    "pillow>=10.0.0,<11.0.0",
    "google-ai-generativelanguage>=0.6.15",
    "google-genai>=1.38.0",
    "pyppmd>=1.2.0",
]
name = "dumpyarabot"
version = "0.1.0"
description = "Bot for AndroidDumps"
readme = "README.md"

[tool.uv]
dev-dependencies = [
    "black<24.0.0,>=23.12.0",
    "mypy<2.0.0,>=1.7.1",
    "isort<6.0.0,>=5.13.2",
    "ruff<1.0.0,>=0.1.8",
    "autoflake<3.0.0,>=2.2.1",
    "pytest>=7.0.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-cov>=4.0.0,<5.0.0",
    "faker>=20.0.0,<21.0.0",
    "responses>=0.24.0,<1.0.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
    "--disable-warnings",
]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow running tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

